requirements.txt
# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
# textract has metadata issues - we'll use alternatives
# textract==1.6.5

# NLP and ML
spacy==3.7.2
nltk==3.8.1
transformers==4.36.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
gensim==4.3.2

# Data Handling
pandas==2.1.4
numpy==1.26.2

# Visualization (optional)
matplotlib==3.8.2
seaborn==0.13.0

# Web Interface (optional)
streamlit==1.29.0
gradio==4.12.0

# Utilities
python-dateutil==2.8.2
regex==2023.12.25

# Download spaCy model after installation:
# python -m spacy download en_core_web_sm